requirements.txt
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.5
django-extra-fields==1.0.0
django-multiselectfield==0.1.8
django-positions==0.6.0
django-rest-swagger==2.2.0
django-timezone-field==3.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
djangorestframework-jwt-refresh-token==0.5
djangorestframework-simplejwt==3.3
Faker==1.0.2
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
openapi-codec==1.3.2
Pillow==5.4.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
PyJWT==1.7.1
pyTelegramBotAPI==3.6.6
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
simplejson==3.16.0
six==1.12.0
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.24.1
